--- a/src/Huffman/Implementation/Blob.cs
+++ b/src/Huffman/Implementation/Blob.cs
@@ -1,47 +1,41 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 
 namespace Huffman.Implementation
 {
     public class Blob
     {
-        private const int ChunkSize = 262_144;
+        private const int InitialBufferSize = 8_000;
 
-        private readonly List<byte[]> _chunks;
-
-        private byte[] _currentChunk;
         private int _bitPosition = -1;
-        private int _bytePosition = ChunkSize - 1;
-        private int _totalLength;
+        private int _bytePosition = -1;
+        private byte[] _bytes;
 
         public Blob()
         {
-            _chunks = new List<byte[]>();
+            _bytes = new byte[InitialBufferSize];
         }
 
         public void Append([NotNull] string bits)
         {
+            var bufferSize = _bytes.Length;
             foreach (var bit in bits)
             {
                 if (_bitPosition == -1)
                 {
                     _bitPosition = 7;
                     _bytePosition++;
-                    _totalLength++;
 
-                    if (_bytePosition == ChunkSize)
+                    if (_bytePosition == bufferSize)
                     {
-                        _currentChunk = new byte[ChunkSize];
-                        _chunks.Add(_currentChunk);
-
-                        _bytePosition = 0;
+                        bufferSize = _bytePosition * 2;
+                        Array.Resize(ref _bytes, bufferSize);
                     }
                 }
 
                 if (bit == '1')
                 {
-                    _currentChunk[_bytePosition] = (byte) (_currentChunk[_bytePosition] | (byte) (1 << _bitPosition));
+                    _bytes[_bytePosition] = (byte) (_bytes[_bytePosition] | (byte) (1 << _bitPosition));
                 }
 
                 _bitPosition--;
@@ -50,22 +44,7 @@ namespace Huffman.Implementation
 
         public byte[] ToByteArray()
         {
-            var buffer = new byte[_totalLength];
-
-            var destinationOffset = 0;
-
-            var toCopy = _totalLength;
-
-            foreach (var chunk in _chunks)
-            {
-                Buffer.BlockCopy(chunk, 0, buffer, destinationOffset, Math.Min(ChunkSize, toCopy));
-
-                destinationOffset += ChunkSize;
-
-                toCopy -= ChunkSize;
-            }
-
-            return buffer;
+            return _bytes[..(_bytePosition+1)];
         }
     }
 }