## Before

|                     Method |      Mean |    Error |   StdDev |
|--------------------------- |----------:|---------:|---------:|
|    Compress_Les_Misérables | 129.10 ms | 1.480 ms | 1.236 ms |
| Compress_War_of_the_Worlds |  14.35 ms | 0.231 ms | 0.247 ms |

|                       Method |      Mean |     Error |    StdDev |
|----------------------------- |----------:|----------:|----------:|
|    Decompress_Les_Misérables | 95.236 ms | 1.5021 ms | 1.3316 ms |
| Decompress_War_of_the_Worlds |  9.778 ms | 0.1480 ms | 0.1155 ms |

## After

|                     Method |      Mean |    Error |   StdDev |
|--------------------------- |----------:|---------:|---------:|
|    Compress_Les_Misérables | 112.91 ms | 2.216 ms | 1.730 ms |
| Compress_War_of_the_Worlds |  14.10 ms | 0.145 ms | 0.129 ms |

|                       Method |      Mean |     Error |    StdDev |    Median |
|----------------------------- |----------:|----------:|----------:|----------:|
|    Decompress_Les_Misérables | 94.826 ms | 1.8838 ms | 4.5134 ms | 93.654 ms |
| Decompress_War_of_the_Worlds |  9.969 ms | 0.1972 ms | 0.3705 ms |  9.816 ms |