@stevehjohn 2021-10-31
Nice @slang25. Looks like a near 10% improvement on compression.

Unfortunately, it seems to have slowed down decompression. Will figure out which bit did that and then get this merged.

On my machine:

### Before

|                     Method |      Mean |    Error |   StdDev |
|--------------------------- |----------:|---------:|---------:|
|    Compress_Les_Misérables | 103.56 ms | 1.155 ms | 0.965 ms |
| Compress_War_of_the_Worlds |  12.75 ms | 0.179 ms | 0.168 ms |

|                       Method |      Mean |     Error |    StdDev |
|----------------------------- |----------:|----------:|----------:|
|    Decompress_Les_Misérables | 81.255 ms | 1.1592 ms | 1.0843 ms |
| Decompress_War_of_the_Worlds |  9.855 ms | 0.0753 ms | 0.0668 ms |

### After

|                     Method |     Mean |    Error |   StdDev |
|--------------------------- |---------:|---------:|---------:|
|    Compress_Les_Misérables | 93.58 ms | 1.104 ms | 1.033 ms |
| Compress_War_of_the_Worlds | 10.54 ms | 0.145 ms | 0.136 ms |

|                       Method |      Mean |     Error |    StdDev |
|----------------------------- |----------:|----------:|----------:|
|    Decompress_Les_Misérables | 90.998 ms | 0.8235 ms | 0.7703 ms |
| Decompress_War_of_the_Worlds |  9.682 ms | 0.0376 ms | 0.0333 ms |