@stevehjohn 2021-10-31
Must be the `ArrayPool` as that's the only change you made for decompression (apart from the avoid double read, which does register as an improvement).