@stevehjohn 2021-10-31
Also (and this is just being really nitpicky now), the static members of the `PriorityQueue` would prevent it from being used on different types of object. In this project, that'll never be the case. Just sayin' that were `PriorityQueue` in a library to be used by different projects (which it won't), it could be an issue.