@stevehjohn 2021-10-31
Lastly, there's a R# `Possible 'null' assignment to non-nullable entity` on those static delegates. My anal-ness won't let that in, sorry! 🤣
Thanks for getting involved - great improvement on compression where I thought there were no more to be made.